backend/pyproject.toml
[project]
name = "eduhu-assistant"
version = "0.1.0"
description = "AI Teaching Assistant with Memory"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-ai>=0.2",
    "anthropic>=0.52",
    "httpx>=0.28",
    "logfire>=3",
    "python-multipart>=0.0.20",
    "pdfplumber>=0.11",
    "python-docx>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.9",
]
